Added NotOverwritable attribute detection to writer

+++ src/Writer.php
@@ -4,16 +4,24 @@ declare(strict_types=1);
 
 namespace Kynx\Mezzio\OpenApiGenerator;
 
+use Kynx\Mezzio\OpenApi\Attribute\NotOverwritable;
+use Nette\InvalidStateException;
+use Nette\PhpGenerator\Attribute;
 use Nette\PhpGenerator\ClassLike;
 use Nette\PhpGenerator\PhpFile;
 use Nette\PhpGenerator\PhpNamespace;
 use Nette\PhpGenerator\Printer;
 use Nette\PhpGenerator\PsrPrinter;
 
+use function array_map;
 use function assert;
 use function current;
+use function file_exists;
+use function file_get_contents;
 use function file_put_contents;
+use function in_array;
 use function is_dir;
+use function is_file;
 use function ltrim;
 use function mkdir;
 use function preg_replace;
@@ -45,9 +53,34 @@ final class Writer implements WriterInterface
         }
 
         $path = $directory . '/' . $this->getFileName($file);
+        if (! $this->canOverWrite($path)) {
+            return;
+        }
+
         if (! @file_put_contents($path, $this->printer->printFile($file))) {
             throw WriterException::cannotWriteFile($path);
         }
+    }
+
+    private function canOverWrite(string $path): bool
+    {
+        if (! (file_exists($path) && is_file($path))) {
+            return true;
+        }
+
+        try {
+            $file = PhpFile::fromCode(file_get_contents($path));
+        } catch (InvalidStateException) {
+            return true;
+        }
+
+        $class = current($file->getClasses());
+        if (! $class instanceof ClassLike) {
+            return true;
+        }
+
+        $attributes = array_map(fn (Attribute $attribute): string => $attribute->getName(), $class->getAttributes());
+        return ! in_array(NotOverwritable::class, $attributes);
     }
 
     private function getDirectory(PhpFile $file): string
